Issue #151 - Add support for CSS/HTML inside typescript files
+ Add test for .jsx/.tsx LS association
+ Extract Angular test into separate class
