+++ org.eclipse.wildwebdeveloper.tests/META-INF/MANIFEST.MF
@@ -16,4 +16,6 @@ Require-Bundle: org.eclipse.wildwebdeveloper,
  org.eclipse.ui.tests.harness,
  org.eclipse.ui,
  org.eclipse.lsp4e,
- org.eclipse.wildwebdeveloper.xml;bundle-version="0.1.0"
+ org.eclipse.wildwebdeveloper.xml;bundle-version="0.1.0",
+ org.eclipse.lsp4j,
+ org.eclipse.lsp4j.jsonrpc
