+++ org.eclipse.wildwebdeveloper.tests/build.properties
@@ -1,4 +1,5 @@
 source.. = src/
 output.. = bin/
 bin.includes = META-INF/,\
+               testProjects/,\
                .
