+++ org.eclipse.wildwebdeveloper.tests/src/org/eclipse/wildwebdeveloper/tests/TestLanguageServers.java
@@ -14,28 +14,16 @@ package org.eclipse.wildwebdeveloper.tests;
 
 import static org.junit.Assert.assertTrue;
 
-import java.io.BufferedReader;
 import java.io.ByteArrayInputStream;
-import java.io.File;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.net.URL;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.nio.file.StandardCopyOption;
-import java.util.stream.Collectors;
 
 import org.eclipse.core.resources.IFile;
-import org.eclipse.core.resources.IMarker;
 import org.eclipse.core.resources.IProject;
 import org.eclipse.core.resources.IResource;
 import org.eclipse.core.resources.ResourcesPlugin;
 import org.eclipse.core.runtime.CoreException;
-import org.eclipse.core.runtime.FileLocator;
-import org.eclipse.core.runtime.NullProgressMonitor;
-import org.eclipse.core.runtime.Path;
-import org.eclipse.core.runtime.Platform;
 import org.eclipse.lsp4e.LanguageServerPlugin;
+import org.eclipse.ui.IViewReference;
+import org.eclipse.ui.IWorkbenchPage;
 import org.eclipse.ui.PlatformUI;
 import org.eclipse.ui.ide.IDE;
 import org.eclipse.ui.tests.harness.util.DisplayHelper;
@@ -65,6 +53,10 @@ public class TestLanguageServers {
 		this.project = ResourcesPlugin.getWorkspace().getRoot().getProject(getClass().getName() + System.nanoTime());
 		project.create(null);
 		project.open(null);
+		IWorkbenchPage activePage = PlatformUI.getWorkbench().getActiveWorkbenchWindow().getActivePage();
+		for (IViewReference ref : activePage.getViewReferences()) {
+			activePage.hideView(ref);
+		}
 	}
 
 	@After
@@ -89,7 +81,7 @@ public class TestLanguageServers {
 					return false;
 				}
 			}
-		}.waitForCondition(PlatformUI.getWorkbench().getDisplay(), 3000));
+		}.waitForCondition(PlatformUI.getWorkbench().getDisplay(), 5000));
 	}
 
 	@Test
@@ -108,7 +100,7 @@ public class TestLanguageServers {
 					return false;
 				}
 			}
-		}.waitForCondition(PlatformUI.getWorkbench().getDisplay(), 3000));
+		}.waitForCondition(PlatformUI.getWorkbench().getDisplay(), 5000));
 	}
 
 	@Test
@@ -127,7 +119,7 @@ public class TestLanguageServers {
 					return false;
 				}
 			}
-		}.waitForCondition(PlatformUI.getWorkbench().getDisplay(), 3000));
+		}.waitForCondition(PlatformUI.getWorkbench().getDisplay(), 5000));
 	}
 
 	@Test
@@ -146,7 +138,7 @@ public class TestLanguageServers {
 					return false;
 				}
 			}
-		}.waitForCondition(PlatformUI.getWorkbench().getDisplay(), 3000));
+		}.waitForCondition(PlatformUI.getWorkbench().getDisplay(), 5000));
 	}
 
 	@Test
@@ -165,7 +157,7 @@ public class TestLanguageServers {
 					return false;
 				}
 			}
-		}.waitForCondition(PlatformUI.getWorkbench().getDisplay(), 3000));
+		}.waitForCondition(PlatformUI.getWorkbench().getDisplay(), 5000));
 	}
 
 	@Test
@@ -184,7 +176,7 @@ public class TestLanguageServers {
 					return false;
 				}
 			}
-		}.waitForCondition(PlatformUI.getWorkbench().getDisplay(), 3000));
+		}.waitForCondition(PlatformUI.getWorkbench().getDisplay(), 5000));
 	}
 
 	@Test
@@ -203,7 +195,7 @@ public class TestLanguageServers {
 					return false;
 				}
 			}
-		}.waitForCondition(PlatformUI.getWorkbench().getDisplay(), 3000));
+		}.waitForCondition(PlatformUI.getWorkbench().getDisplay(), 5000));
 	}
 
 	@Test
@@ -222,7 +214,7 @@ public class TestLanguageServers {
 					return false;
 				}
 			}
-		}.waitForCondition(PlatformUI.getWorkbench().getDisplay(), 3000));
+		}.waitForCondition(PlatformUI.getWorkbench().getDisplay(), 5000));
 	}
 
 	@Test
@@ -241,7 +233,7 @@ public class TestLanguageServers {
 					return false;
 				}
 			}
-		}.waitForCondition(PlatformUI.getWorkbench().getDisplay(), 3000));
+		}.waitForCondition(PlatformUI.getWorkbench().getDisplay(), 5000));
 	}
 
 	@Test
@@ -260,106 +252,44 @@ public class TestLanguageServers {
 					return false;
 				}
 			}
-		}.waitForCondition(PlatformUI.getWorkbench().getDisplay(), 3000));
+		}.waitForCondition(PlatformUI.getWorkbench().getDisplay(), 5000));
 	}
 
 	@Test
-	public void testAngularTSFile() throws Exception {
-		URL url = FileLocator.find(Platform.getBundle("org.eclipse.wildwebdeveloper.tests"),
-				Path.fromPortableString("testProjects/angular-app"), null);
-		url = FileLocator.toFileURL(url);
-		File folder = new File(url.getFile());
-		if (folder.exists()) {
-			
-			java.nio.file.Path sourceFolder = folder.toPath();
-			java.nio.file.Path destFolder = project.getLocation().toFile().toPath();
-			
-			Files.walk(sourceFolder).forEach(source -> {
+	public void testJSXFile() throws Exception {
+		final IFile file = project.getFile("blah.jsx");
+		file.create(new ByteArrayInputStream("ERROR".getBytes()), true, null);
+		ITextEditor editor = (ITextEditor) IDE.openEditor(PlatformUI.getWorkbench().getActiveWorkbenchWindow().getActivePage(), file);
+		editor.getDocumentProvider().getDocument(editor.getEditorInput()).set("a<");
+		assertTrue("Diagnostic not published", new DisplayHelper() {
+			@Override
+			protected boolean condition() {
 				try {
-					Files.copy(source, destFolder.resolve(sourceFolder.relativize(source)), StandardCopyOption.REPLACE_EXISTING);
-				} catch (IOException e) {
-					e.printStackTrace();
-				}	
-			});
-			
-			Process process = new ProcessBuilder(getNpmLocation(), "install", "--no-bin-links", "--ignore-scripts")
-					.directory(project.getLocation().toFile()).start();
-			while (process.isAlive())
-				;
-			project.refreshLocal(IResource.DEPTH_INFINITE, new NullProgressMonitor());
-
-			IFile file = project.getFolder("src").getFolder("app").getFile("app.component.ts.content");
-			String fileContent = new BufferedReader(new InputStreamReader(file.getContents())).lines()
-					.collect(Collectors.joining("\n"));
-
-			IFile appComponentFile = project.getFolder("src").getFolder("app").getFile("app.component.ts");
-			appComponentFile.create(new ByteArrayInputStream("".getBytes()), true, null);
-			assertTrue("Diagnostic published on empty file", new DisplayHelper() {
-				@Override
-				protected boolean condition() {
-					try {
-						return appComponentFile.findMarkers("org.eclipse.lsp4e.diagnostic", true,
-								IResource.DEPTH_ZERO).length == 0;
-					} catch (CoreException e) {
-						return false;
-					}
-				}
-			}.waitForCondition(PlatformUI.getWorkbench().getDisplay(), 3000));
-
-			ITextEditor editor = (ITextEditor) IDE
-					.openEditor(PlatformUI.getWorkbench().getActiveWorkbenchWindow().getActivePage(), appComponentFile);
-			editor.getDocumentProvider().getDocument(editor.getEditorInput()).set(fileContent);
-			assertTrue("Diagnostic not published", new DisplayHelper() {
-				@Override
-				protected boolean condition() {
-					try {
-						return appComponentFile.findMarkers("org.eclipse.lsp4e.diagnostic", true,
-								IResource.DEPTH_ZERO).length != 0;
-					} catch (CoreException e) {
-						return false;
-					}
-				}
-			}.waitForCondition(PlatformUI.getWorkbench().getDisplay(), 50000));
-
-			IMarker[] markers = appComponentFile.findMarkers("org.eclipse.lsp4e.diagnostic", true,
-					IResource.DEPTH_ZERO);
-			boolean foundError = false;
-			for (IMarker marker : markers) {
-				int lineNumber = marker.getAttribute(IMarker.LINE_NUMBER, -1);
-				if (lineNumber == 6 && marker.getAttribute(IMarker.MESSAGE, "").contains("template")) {
-					foundError = true;
+					return file.findMarkers("org.eclipse.lsp4e.diagnostic", true, IResource.DEPTH_ZERO).length != 0;
+				} catch (CoreException e) {
+					return false;
 				}
 			}
-			assertTrue("No error found in line 6", foundError);
-		}
+		}.waitForCondition(PlatformUI.getWorkbench().getDisplay(), 5000));
 	}
 
-	public static String getNpmLocation() {
-		String res = "/path/to/npm";
-		String[] command = new String[] { "/bin/bash", "-c", "which npm" };
-		if (Platform.getOS().equals(Platform.OS_WIN32)) {
-			command = new String[] { "cmd", "/c", "where npm" };
-		}
-		BufferedReader reader = null;
-		try {
-			Process p = Runtime.getRuntime().exec(command);
-			reader = new BufferedReader(new InputStreamReader(p.getInputStream()));
-			res = reader.readLine();
-		} catch (IOException e) {
-			return Platform.getOS().equals(Platform.OS_WIN32) ? "npm.cmd" : "npm";
-		}
-
-		// Try default install path as last resort
-		if (res == null && Platform.getOS().equals(Platform.OS_MACOSX)) {
-			res = "/usr/local/bin/npm";
-		} else if (res == null && Platform.getOS().equals(Platform.OS_LINUX)) {
-			res = "/usr/bin/npm";
-		}
-
-		if (res != null && Files.exists(Paths.get(res))) {
-			return res;
-		}
-		return Platform.getOS().equals(Platform.OS_WIN32) ? "npm.cmd" : "npm";
+	@Test
+	public void testTSXFile() throws Exception {
+		final IFile file = project.getFile("blah.tsx");
+		file.create(new ByteArrayInputStream("ERROR".getBytes()), true, null);
+		ITextEditor editor = (ITextEditor) IDE.openEditor(PlatformUI.getWorkbench().getActiveWorkbenchWindow().getActivePage(), file);
+		editor.getDocumentProvider().getDocument(editor.getEditorInput()).set("FAIL");
+		assertTrue("Diagnostic not published", new DisplayHelper() {
+			@Override
+			protected boolean condition() {
+				try {
+					return file.findMarkers("org.eclipse.lsp4e.diagnostic", true, IResource.DEPTH_ZERO).length != 0;
+				} catch (CoreException e) {
+					return false;
+				}
+			}
+		}.waitForCondition(PlatformUI.getWorkbench().getDisplay(), 5000));
 	}
+
 
 }
