+++ org.eclipse.wildwebdeveloper.tests/testProjects/htmlIn.tsx
@@ -1,0 +1,1 @@
+var a = html`<button>Hello</button>`;
