+++ org.eclipse.wildwebdeveloper.xml/pom.xml
@@ -26,7 +26,7 @@
 			<plugin>
 				<groupId>com.googlecode.maven-download-plugin</groupId>
 				<artifactId>download-maven-plugin</artifactId>
-				<version>1.3.0</version>
+				<version>1.4.1</version>
 				<executions>
 					<execution>
 						<id>fetch-xml</id>
