+++ org.eclipse.wildwebdeveloper/grammars/typescript/MarkdownDocumentationInjection.tmLanguage.json
@@ -1,0 +1,29 @@
+{
+	"injectionSelector": "L:comment.block.documentation",
+	"patterns": [
+		{
+			"include": "#markdown-comment"
+		}
+	],
+	"repository": {
+		"markdown-comment": {
+			"begin": "(?<=/\\*\\*)([^*]|\\*(?!/))*$",
+			"while": "(^|\\G)\\s*\\*(?!/)(?=([^*]|[*](?!/))*$)",
+			"patterns": [
+				{
+					"include": "text.html.markdown#fenced_code_block"
+				},
+				{
+					"include": "text.html.markdown#lists"
+				},
+				{
+					"include": "source.ts#docblock"
+				},
+				{
+					"include": "text.html.markdown#inline"
+				}
+			]
+		}
+	},
+	"scopeName": "documentation.markdown.injection"
+}
