+++ org.eclipse.wildwebdeveloper/grammars/typescript/Readme.md
@@ -1,0 +1,2 @@
+Those files are taken from VSCode.
+Do not modify them here. Contribute fix to VSCode and once patch is merged in VSCode, make Wild Web Developer include newer grammars from VSCode.
