+++ org.eclipse.wildwebdeveloper/language-servers/package.json
@@ -14,6 +14,8 @@
     "vscode-css-languageserver": "./vscode-css-languageserver-1.0.0.tgz",
     "vscode-html-languageserver": "./vscode-html-languageserver-1.0.0.tgz",
     "vscode-json-languageserver": "./vscode-json-languageserver-1.0.1.tgz",
-    "yaml-language-server": "0.3.2"
+    "yaml-language-server": "0.3.2",
+    "typescript-plugin-css-modules": "1.0.5",
+    "typescript-lit-html-plugin": "0.9.0"
   }
 }
