+++ org.eclipse.wildwebdeveloper/src/org/eclipse/wildwebdeveloper/jsts/JSTSLanguageServer.java
@@ -47,12 +47,12 @@ public class JSTSLanguageServer extends ProcessStreamConnectionProvider {
 	@Override
 	public Object getInitializationOptions(URI rootUri) {
 		Map<String, Object> options = new HashMap<>();
-		List<Map<String, String>> plugins = new ArrayList<>();
+		List<TypeScriptPlugin> plugins = new ArrayList<>();
 		try {
-			URL url = FileLocator.toFileURL(getClass().getResource("/language-servers/node_modules/@angular/language-service"));
-			plugins.add(new TypeScriptPlugin("@angular/language-service", url.getPath()).toMap());
-			options.put("plugins", plugins);
-			
+			plugins.add(new TypeScriptPlugin("@angular/language-service"));
+			plugins.add(new TypeScriptPlugin("typescript-plugin-css-modules"));
+			plugins.add(new TypeScriptPlugin("typescript-lit-html-plugin"));
+			options.put("plugins", plugins.stream().map(TypeScriptPlugin::toMap).toArray());
 		} catch (IOException e) {
 			Activator.getDefault().getLog().log(new Status(IStatus.ERROR, Activator.getDefault().getBundle().getSymbolicName(), e.getMessage(), e));
 		}
