+++ target-platform/target-platform.target
@@ -9,7 +9,7 @@
 	<unit id="org.mockito" version="0.0.0"/>
 </location>
 <location includeAllPlatforms="false" includeConfigurePhase="false" includeMode="planner" includeSource="true" type="InstallableUnit">
-	<repository location="http://download.eclipse.org/staging/2019-03/"/>
+	<repository location="http://download.eclipse.org/releases/2019-03/"/>
 	<unit id="tm-feature.feature.group" version="0.0.0"/>
 	<unit id="org.objectweb.asm" version="0.0.0"/>
 	<unit id="org.eclipse.mylyn.wikitext_feature.feature.group" version="0.0.0"/>
